feat: add useLSSavedSearch hook
